outlining kernels
