+++ apps/c/airfoil/airfoil_tutorial/final/adt_calc.h
@@ -1,0 +1,28 @@
+//outlined elemental kernel - adt_calc
+inline void adt_calc(const double *x1, const double *x2, const double *x3,
+                     const double *x4, const double *q, double *adt) {
+  double dx, dy, ri, u, v, c;
+
+  ri = 1.0f / q[0];
+  u = ri * q[1];
+  v = ri * q[2];
+  c = sqrt(gam * gm1 * (ri * q[3] - 0.5f * (u * u + v * v)));
+
+  dx = x2[0] - x1[0];
+  dy = x2[1] - x1[1];
+  *adt = fabs(u * dy - v * dx) + c * sqrt(dx * dx + dy * dy);
+
+  dx = x3[0] - x2[0];
+  dy = x3[1] - x2[1];
+  *adt += fabs(u * dy - v * dx) + c * sqrt(dx * dx + dy * dy);
+
+  dx = x4[0] - x3[0];
+  dy = x4[1] - x3[1];
+  *adt += fabs(u * dy - v * dx) + c * sqrt(dx * dx + dy * dy);
+
+  dx = x1[0] - x4[0];
+  dy = x1[1] - x4[1];
+  *adt += fabs(u * dy - v * dx) + c * sqrt(dx * dx + dy * dy);
+
+  *adt = (*adt) / cfl;
+}
