+++ apps/c/airfoil/airfoil_tutorial/final/save_soln.h
@@ -1,0 +1,5 @@
+//outlined elemental kernel - save_soln
+inline void save_soln(const double *q, double *qold) {
+  for (int n = 0; n < 4; n++)
+    qold[n] = q[n];
+}
